FIAPP/FIAPP/requirements.txt
Flask>=2.0.0
firebase-admin>=6.0.0
python-dotenv>=0.19.0
requests>=2.25.0
groq>=0.1.0
